build: 修正 msvc 编译错误和警告

+++ gtest/TestRandom.cpp
@@ -14,7 +14,7 @@ TEST(TestRandom, STD) {
     ASSERT_NE(file , nullptr);
     auto writer = sese::CSVWriter(file.get(), ',', false);
     for (int i = 0 ; i < 300000; i++) {
-        std::default_random_engine engine(device());
+        std::default_random_engine engine((unsigned int)device());
         writer.write({std::to_string(i), std::to_string(distribution(engine))});
     }
     file->close();
